requirements.txt
asgiref==1.0.0
autobahn==0.17.1
cffi==1.9.1
channels==1.0.2
constantly==15.1.0
cryptography==1.7.1
daphne==1.0.1
Django==1.10.5
django-guardian==1.4.6
idna==2.2
incremental==16.10.1
paramiko==2.1.1
pyasn1==0.1.9
pycparser==2.17
six==1.10.0
Twisted==16.6.0
txaio==2.6.0
websocket-client==0.40.0
zope.interface==4.3.3
pywin32==220
